Add invalid configuration check for ITS_ENCRYPTION
ITS_ENCRYPTION cannot be used when PSA_FRAMEWORK_HAS_MM_IOVEC
is also enabled.

ITS_ENCRYPTION requires buffers which are only available when
memory mapped vectors are disabled.

This miscoconfiguration fails to compile ITS with a compilation
error which is not very helpful. This makes it more clear what is
the issue.

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>
Change-Id: I5de12aea8ebf76f5ce258fc16f2d1b9d053174cd