secure_fw: agent_psa_connect returns a status not a handle
agent_psa_connect() and its respective SPM handler perform a connection
setup but the actual handle that's going to be returned to the initial
caller is saved into the connection.

Therefore, the correct type for the returned value to agent_psa_connect
is psa_status_t.

Alongside, remove unnecessary castings.

Signed-off-by: Nicola Mazzucato <nicola.mazzucato@arm.com>
Change-Id: I9a2d7d7a3b5834240433a84746e0cf28e386f42b