SPM: psa_api: Fix return value to client when psa_connect fails
When the SPM or RoT refuse the connection or cannot make the connection,
psa_connect can return two errors: either
PSA_ERROR_CONNECTION_REFUSED or PSA_ERROR_CONNECTION_BUSY.

In the first case, in tfm_spm_partition_psa_reply the connection is set
to be deleted immediately, while in the latter the connection is not
deleted at all.

This leaves us into two unwanted scenarios:
 a. deleting the connection immediately would cause its fields to be
    filled with some predefined values - in debug builds - to catch
    use-after-free cases. This in turn is reflected in the returned
    value to the caller, instead of the expected PSA error value.
 b. the connection is not deleted and never recovered so eventually the
    allocator will run out of free connections.

Fix both scenarios above by setting the connection handle status to
TFM_HANDLE_STATUS_TO_FREE which will allow SPM to correctly pick up the
corresponding PSA_ERROR and return it to the caller. The connection
will then be freed immediately if the backend is SFN or later in IPC
model.

Signed-off-by: Nicola Mazzucato <nicola.mazzucato@arm.com>
Change-Id: I7f87ec004f076d71640c7a339d44a90f7a015347