platform: nordic_nrf: Use nordic_hal tag instead of commit hash
Using a commit hash and shallow mode causes issues when fetching the
HAL. This replaces the commit hash with a git tag which should work.
The git tag points to a later version of the nordic_hal compared to
the commit hash but this should not cause any issues since the
changelog doesn't report any breaking changes for this vesion.

Change-Id: I1c9d1ba6fd9f3e9f3da4708e6a1324defb8d7815
Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>