SPM: tfm_multi_core: Record and clear mailbox IRQ via multi_core API
With the recent introduction of the NS_AGENT_MBOX_PROCESS_NEW_MSG
auxiliary service, the interrupt corresponding to the mailbox message
needs to be managed.
Thus, add a basic table in multi-core implementation to track the
asserted interrupts that need to be cleared. This table is accessed via
an API defined in multi-core and generated at build time with a size
that's taken from the mailbox manifest file.

Right after the time when the mailbox interrupt handler calls
`spm_handle_interrupt`, the corresponding IRQ number will be stored in
the table by `tfm_multi_core_set_mbox_irq`.
That interrupt will be cleared when once the deferred `process_new_msg`
action has taken place.

Signed-off-by: Nicola Mazzucato <nicola.mazzucato@arm.com>
Change-Id: Ia1886607d1d02ea0137090ddc2fa81066a30fae5
Signed-off-by: Anton Komlev <anton.komlev@arm.com>