SPM: psa_api: Free connection in backend for DISCONNECT
With the ipc backend, in case of psa_close, the delete_connection
boolean is set and the connection is freed right away.
This leaves incorrectly a reference later in the p_replied to a
freed connection.

Instead, the connection needs to be freed after the replied value has
been taken, in the query_state.

Fix the above by setting the status in the connection and prevent the
immediate free-up of the connection.

Signed-off-by: Nicola Mazzucato <nicola.mazzucato@arm.com>
Change-Id: I454f2f30c4e2447ecaf8b9285963d00aeb084081