LIB: Do not calculate string length unnecessarily
Previously, the output string function always calculated the length of
the string given to it, even when the value was known by the calling
function. Optimise the logging library further by passing the string
length to this function and only iterating through the string to
calculate the length when required.

Change-Id: I6d6a1580fccf6385b7b1675277e817ea300e3224
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>