RSE: Support use of CM ROTPKs for DM hash check
Check the DM blob header fields to determine if it is verified using the
public key available in the blob itself (which has the hash in the CM
ROTPK fields).

Calculate a hash of this key and compare it to that in
the OTP (which will have been previously been provisioned) before
passing the key to the higher level functions which will use it to check
the signature of the blob.

Change-Id: I5cdccd67d3f023746645818c7fda055f37221a87
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>