platform: nordic_nrf: Add nRF54L15 initial support
Before this there was only support for building the nRF54L15.
This updates the nRF54L15 files to add an initial support for
running TF-M.

This is NOT full upstream support. There are important limitations:
- Hardware crypto acceleration with Cracen is not supported
- Random number generation with Cracen is not supported
- The tests suites PSA arch tests and TF-M regression tests
  are not tested
- BL2 is not supported
- Some soc related configurations are not supported and they rely
  on hard-coded values (check nordicsemi_nrf54l_init for more info)

This was tested using some basic Zephyr samples.

This change updates the flash_layout.h and the region_maps.h headers
with a layout that makes sense for the nRF54L15. Only one layout
is supported at the moment.

The RTE_Device.h is modified to use the UART20 and UART30 ports
in order to support UART serial output.

The function nordicsemi_nrf54l_init was copied from the soc.c in
Zephyr and was simplified to support one single configuration.
Later it can be considered to add configurability if needed.

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>
Change-Id: I0369987cc1f4e7994078202e9d1bcb53c6372281
(cherry picked from commit 01b4551f0b191c3238e6371bbbe7e7aeb8b64636)