RSE: Add chained provisioning flow
Add code and CMake configuration to support the chained provisioning
flow. In this flow the main provisioning code first loads a 'chained'
blob which contains a key which will be used to verify the subsequent
'main' DM blob. In order to facilitate this, we add a new function to
the ROTPK file which allows the 'chained' blob to setup the key which
will then be used when the next blob is loaded (and the key type is set
to PREVIOUS_BLOB in the header). Note that the set key must persist across
reboots and therefore it is placed in the new .persistent_data section.

Change-Id: Ic71898b8ee9fb68a927c239a7acd9dde4509e0f8
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>