Tools: Correct key_derivation.py argument parsing
Previously the hkdf_hash_alg argument was given a default value which
meant that the parse_args function always returned a non-empty
dictionary, even when the kdf_alg argument was not provided. This
dictionary is used in calling scripts to check if the script invocation
has requested key derivation. Therefore refactor the script to move the
default argument to the function.

Change-Id: I4c1e005634d498456675b85f0e1cb0e941d353c1
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>