RSE: Fix unaligned pointer warnings
Fix warnings of the following type:
warning: taking address of packed member of
'struct rse_otp_cm_area_t' may result in an
unaligned pointer value [-Waddress-of-packed-member]

Achieve this by changing the function signature of
rse_check_zero_bit_count to pass a value instead of pointer for
expected_zero_count. This eliminates the need for expected_zero_count to
be 4 byte aligned.

Change-Id: I8d3b4578d13ab36d14f624564cd08c89013cbcfc