RSE: Correct provisioning message encryption args
The provisioning_message_config.py takes in arguments to determine
whether or not to encrypt the code and secret values. The arguments are
defined as enum arguments with a default value, but are evaluated in the
script as a bool (so always evaluate to true). Change the use in the
script to instead check the value of the strings instead.

Change-Id: I5703c842580b1afa015b968ecf1734db4551d3f4
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>