RSE: Add carve-out for OTP emulation
Add a memory carve-out to support OTP emulation in SRAM. This patch
specifies a new reserved memory section which is reserved from each of
the RSE stages. Note that is it only actually added to the BL1_1 linker
script as that is the only boot stage which will directly make use of
this memory (by writing its address into a OTP emulation register).

This patch also adds additional checks to ensure that the memory map is
more robust (ensuring that none of the calculations are invalid and that
everything fits in the SRAM). None of these checks are exhaustive, but
hopefully means we are more likely to catch bugs in the memory map
sizing.

Change-Id: Ifc5a0d13b1830b579134c120a8aa221667c47cfe
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>