BL2: flash_map: Fix MISRA warnings in flash_area_read()
MISRA warning MC3R1.R9.1 states: "The value of an object with automatic
storage duration shall not be read before it has been set". The caution
was trigger due to the repeating `if(remaining_len)` statement that
caused assignment of `item_number` to be in different scopes.

Fix code format for concluding `return 0;` to match rest of the
codebase.

Change-Id: I4ee6d2b505980d53be50eb4915254c214eb1f039
Signed-off-by: Mudit Sharma <mudit.sharma@arm.com>