RSE: Tests: Refine FVP limitation workaround
The previous check allowed for tests which required SP enabled to be
false to run, even when the hardware reported the SP to be enabled
(because of the overriding of the sp_enabled variable). This lead to
some tests being run when they should not have.

Instead, refine the check to not run any tests which require a specific
SP state when the hardware is in an invalid state.

Change-Id: Icfdd5373d85b5e1204562580e3861cf4fae60db5
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>