FWU: Fix MISRA warnings for rule 9.1
MISRA rule 9.1 states: "The value of an object with automatic storage
duration shall not be read before it has been set".

Looking at the flow of the code (from top to bottom) `blk_sz` seems
to be used uninitialized during setting the value of `off` in the
for loop. However, that is not really the case, as we only use the
`blk_sz` after the first loop, in which we set the value of
`blk_sz`. Initializing the value of `blk_sz` with this commit simply
helps with code readability and resolve a MISRA caution.

Change-Id: Icbca32920d9af8359fe2a45b7d3dc3dd0ff5a5ad
Signed-off-by: Mudit Sharma <mudit.sharma@arm.com>