CORE: Fix MISRA warnings for rule 9.1
MISRA rule 9.1 states: "The value of an object with automatic storage
duration shall not be read before it has been set".

Looking at the flow of the code (from top to bottom) `next_tlv_offset`
seems to be used uninitialized during setting the value of `offset` in
the for loop. However, that is not really the case, as we only use the
`next_tlv_offset` after the first loop, in which we set the value
of `next_tlv_offset`. Initializing the value of `next_tlv_offset` with
this commit simply helps with code readability.

Change-Id: Id9767af1b8b8d8e36145006b1285754f4c1f13af
Signed-off-by: Mudit Sharma <mudit.sharma@arm.com>