Platform: Corstone1000: Fix .data loading in BL1
The .data section's LMA is in FLASH, and it is loaded into RAM at
runtime. Because of this, the .data section's size must be taken into
consideration when checking for FLASH overflow during link-time.

This is not checked by GCC's linker when the LMA is set to __etext, so
the linker scripts had to be updated to explicitly state that the .data
section is loaded into FLASH.

After this commit, the linker properly checks for FLASH overflow and
also accounts for the .data section's size.
The [1] was used as a reference for this change.

The BL1_1_CODE_SIZE had to be increased because it was previously insufficient.

[1] https://github.com/ARM-software/CMSIS_6/blob/v6.1.0/CMSIS/CoreValidation/Layer/Target/CM55/RTE/Device/ARMCM55/gcc_linker_script.ld

Change-Id: Ia82faa49023711c5518ee4cd3e7ab7141f5d507c
Signed-off-by: Bence Balogh <bence.balogh@arm.com>