CC3XX: modify rng config check to avoid static analysis error
cc3xx_rng_rosc_id_t is defined as a packed enum starting with 0, meaning
the compiler *MAY* interpret it as unsigned char. This results in
"rosc_id < CC3XX_RNG_ROSC_ID_0" being always false.

However the compiler may choose to interpret it as signed char, then the
check may be either true or false.

To avoid the result being compiler-dependent, adapt the code to check
that the id is within the bounds instead.

Change-Id: I81068f7edb27a0aadee20de10e337810637555f0
Signed-off-by: Amjad Ouled-Ameur <amjad.ouled-ameur@arm.com>