RSE: Calculate ROTPK field sizes
Use the sizes of the ROTPK hashes contained in the ROTPKs to determine
the required sizes of the fields. This patch also adds a new CMake
configuration variable which specifies the hash algorithm to use with
the DM signing key in the case that this is embedded in the blob.

Change-Id: I86a635377f4595c38852a64106a812cd455ab384
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>