nordic_nrf: Fix UART and FLASH build issues
The nordic platforms enable SECURE_UART1 by default,
enabling this without TFM_SPM_LOG_RAW_ENABLED doesn't make sense
since it will configure a UART as secure and then it will not use
it. This also causes a building issue since the code expects both
these two options to be enabled.

Also fixes a building issue related to flash. The cmsis flash driver
is used by both PLATFORM_DEFAULT_OTP and byITS but it was only
included when ITS was build. So it was failing for builds with
PLATFORM_DEFAULT_OTP and without ITS

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>
Change-Id: I4f8f75ec2a2817634c190ebb844b9be9cc7f1a45