platform: nordic_nrf: Selectively enable secure UART1
Enable the secure UART only when the TF-M logging is
enabled.

This makes sure that the the UART1 is not configured
as secure when there is no logging enabled for TF-M.

This works better with the Zephyr intergration since
the TF-M logging level option exists in Zephyr.

Change-Id: I65c8d80143c95fc22963567cc4b2115b9de7f4e1
Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>
(cherry picked from commit 5b8b25e6891e6b4b3573a377b533efb8ed88d7ff)