platform: nordic_nrf: Create a nrf54l folder
Create a nrf54l folder which can be used for other
devices in the series as well.

This is preparation work for the addition of the
nrf54l10.

This also updates the the source and header files
which use ifdefs with nrf54l15 specific defines with
the macro NRF54L_SERIES.

Also removes the executable permissions from the
region_defs.h which was enabled by accident.

Change-Id: I1b2a0876fefef3fb460cf8995b9b802aa089b7aa
Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>
(cherry picked from commit a49f175edadc8b36a01517f8bcc746d05991f1d5)