LIB: Return total printed chars from vprintf
The vprintf implementation in crt_vprintf.c should return the number of
characters printed (as that is the standard signature of vprintf).
Rework the tfm_vprintf_unpriv implementation to count the number of
characters printed to the UART and return them.

Change-Id: I26f9f4efdede685c1e82294859ef3ce1fcf45fcd
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>
(cherry picked from commit b3118302acca474306019d7040e349229d32748b)
Signed-off-by: Stefan Gloor <stefan.gloor@siemens.com>