make psa_can_do_hash() inline
With the update of Mbed TLS to 3.6.3 MBEDTLS_MD_SOME_PSA can get defined
also when only MBEDTLS_PSA_CRYPTO_CLIENT (and not MBEDTLS_PSA_CRYPTO_C)
is defined, which will make MD's md_can_use_psa() call psa_can_do_hash().

The problem with this is that this gives linker errors when the NS image
tries to call psa_can_do_hash() because it's not made available by TF-M.

This is a quick fix to get things working.
Discussions are ongoing with TF-M and Mbed TLS to see how to fix this
properly.

Signed-off-by: Tomi Fontanilles <tomi.fontanilles@nordicsemi.no>