[zep noup] platform: nordic_nrf: Split the target_cfg for different platforms
This is a noup as the original commit does not apply cleanly.
Attempts to cherry-pick it in cleanly leads to a large amount
of extra commits being included.

Original commit message:
This splits the target_cfg.c file for the different platfoms.
It splits the file into two files one for the nRF54L and one for
the nRF91/nRF53 series.

There is no need for the peripherals to be in the same file as
the configuration code. There is already a header which
provides extern definitions for all the peripherals a
device supports which can be used for for source files
needing to configure peripherals. So this peripheral definitions
are placed in a separate file.

Most of the code is taken unchanged from the target_cfg.c file
and the uneeded ifdefs got removed.

The init_debug function got rewritten without changing the logic
of it.

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>
Change-Id: I19071727061b8952e3abbae20dd182c822234427
(cherry picked from commit 37e8dc47d4eb4aa0c0b905f80be5272aafbd9fa2)