platform: stm: b_u585i_iot02a: Change flash layout
The b_u585i_iot02a board have an external flash memory which is already
enabled inside tf-m. This proposes a new layout to increase the space
for bootloader, S and NS firmware to allow bigger applications to be
developed. The firmware upgrade then will use the first sectors from
external NOR memory as the secundary slots. With this change the
bootloader still have 16k available to handle future needs but keeps the
applications partitions the same.

Change-Id: Ibb1b12cfbd2eca60736fb2d45f7dbea79807381f
Signed-off-by: BUDKE Gerson Fernando <gerson.budke@leica-geosystems.com>