platform: stm32u5xx: Move provision files
The provision files are board dependent but are defined inside common
CMakeLists.txt. This means that the current behaviour do not allow
customize this values for Out of Tree buildings. This moves the
definition to the board CMakeLists.txt to solve the issue.

Change-Id: I6654fa7f61a3f7f33e65ac9b979e9224e009e8e9
Signed-off-by: BUDKE Gerson Fernando <gerson.budke@leica-geosystems.com>

(cherry picked from commit 845319f646659104524c840e76e398150f60ed60)