platform: ext: common: template: Fix FOTA upgrade
The 31755 changes modified profile_definition from 32 to 48 bytes in
the platform/ext/common/template/flash_otp_nv_counters_backend.h file.

This change moved fields that are necessary to perform a an image
upgrade. The issue is visible only when upgrading from an old image
(without the changes) to a new one or vice versa. This means that
TF-M 2.1.1 do not allow upgrade/downgrade to TF-M 2.2.0 or newer.

This revert the change to allow migration from Zephyr 3.7.1 LTS to
newer images like 4.2.0.

More details from the change can be visualized in:
https://review.trustedfirmware.org/c/TF-M/trusted-firmware-m/+/31755

Signed-off-by: BUDKE Gerson Fernando <gerson.budke@leica-geosystems.com>