[zep fromtree] Attest: Clarify the HAL API behaviour for profile_definition
As the OTP might store a reduced representation of the
profile_definition string, the HAL function needs to be
able to translate such reduced representation to the full
string. It is also allowed for the platform HAL to return
an empty string and in such case the profile definition will
have to be determined solely based on build options.

Signed-off-by: Antonio de Angelis <antonio.deangelis@arm.com>
Change-Id: Ia32f4bb30d114623ab5ac0197d33840e0cafc317
(cherry picked from commit d85e8cd9b0ceb7031317048da9d89eba4893e6e3)