[zep fromtree] platform: nordic_nrf: Add memory service header for nRF54LV10A/M20A
This is a continuation of "Add memory service header in common folder"
Adds a copy of the memory_services_range to the common folder
to allow Zephyr builds. A proper solution to this is planned
later to avoid the duplication.

Change-Id: I303c2a24caa3149883ad4ddb592b3ba1be2a61d2
Signed-off-by: Dag Erik Gjørvad <dag.erik.gjorvad@nordicsemi.no>
(cherry picked from commit 2f86a3d4b4d389a5e8c7abbd70541e1718453c62)