platform: Replace picolibc _start instead of using a constructor
Add the remaining picolibc initialization steps to picolibc_startup
and rename it to _start so that it can replace the picolibc
version. This lets us remove the symbol hacks needed to make the
picolibc version happy, including the _set_tls function.

This requires adding '-nostartfiles -u main' to the link options.

Signed-off-by: Keith Packard <keithp@keithp.com>